+++ src/org/pullrequest/android/bookingnative/domain/dao/UserDao.java
@@ -6,7 +6,7 @@ import org.pullrequest.android.bookingnative.domain.model.User;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.support.ConnectionSource;
 
-public interface UserDao extends Dao<User, Integer> {
+public interface UserDao extends Dao<User, Long> {
 
 	ConnectionSource getConnectionSource();
